requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.15.1
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
attrs==25.1.0
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.3
docx==0.2.4
exceptiongroup==1.2.2
fastapi==0.115.11
Flask==3.1.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.4.0
magic-filter==1.0.12
Mako==1.3.9
MarkupSafe==3.0.2
multidict==6.1.0
numpy
pillow==11.1.0
propcache==0.2.1
psutil==6.1.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.1.0
pyTelegramBotAPI==4.26.0
python-dotenv==1.1.1
python-libnmap==0.7.3
repoze.lru==0.7
requests==2.32.3
Routes==2.5.1
scipy==1.13.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.0
telebot==0.0.5
typing_extensions==4.12.2
urllib3==2.3.0
vboxapi==1.0
Werkzeug==3.1.3
yarl==1.18.3
zipp==3.21.0